requirements.txt
Django>=5
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.3
drf-spectacular>=0.27.1
drf-spectacular[sidecar]>=0.27.1
google-genai>=0.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9